requirements.txt
numpy 
pandas
scikit-learn
matplotlib
opencv-python
pytesseract
easyocr
spacy
transformers
torch
python-dotenv
google-generativeai>=0.8.3
tensorflow
nltk
keras
fastapi
uvicorn
Jinja2
python-multipart